+++ src/main/java/core/basesyntax/figures/RightTriangle.java
@@ -15,7 +15,7 @@ public class RightTriangle extends Figure {
 
     @Override
     public double getArea() {
-        return (double) firstLeg * secondLeg * 0.5;
+        return firstLeg * secondLeg * 0.5;
     }
 
     @Override
